Add test for pieces being different types
A test was added that checks pieces can be different types, eg pawns,
knights etc. It is assumed that the Piece class will have a getter that
returns a piece's type, and that the type is initially specified in the
piece's initializer. This meant that other tests using the initialiser
had to be updated to reflect the new number of arguments in the
function.