Refactor Board class
The getter for the spaces member variable was deleted, so spaces is now
private and inaccessible outside of the class.

Various changes were required to other classes accessing the variable.
Primarily this involved using the add_piece and piece_at Board methods
instead.

One slightly more complex change related to board_spec. The first test
checks that a size 8 x 8 board is created. Without access to the spaces
variable, this test had to be rewritten. Instead, the test now attempts
to add pieces that sit outside of the board size, expecting errors to be
raised if that is attempted. The add_piece method in Board had to be
amended to carry out this error checking.